requirements.txt
# 필수 패키지
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
Pillow>=8.2.0
opencv-python>=4.5.2
pycocotools>=2.0.2
tensorboard>=2.5.0

# detectron2 설치는 별도로 진행 필요
# pip install -e git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2